% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{count_na}
\alias{count_na}
\title{Count explicit NA}
\usage{
count_na(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{A bare variable that may contain \code{NA}.}
}
\value{
A tibble contains:
\itemize{
\item the "key" of the \code{tbl_ts}
\item ".from": the starting time point of \code{NA}
\item ".to": the ending time point of \code{NA}
\item ".n": the number of \code{NA} during the time period
}
}
\description{
Count explicit NA
}
\examples{
pedestrian \%>\%
  fill_gaps(.full = TRUE) \%>\%
  count_na(Count)
}
