% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{layer_na_rle}
\alias{layer_na_rle}
\title{Add a layer of \code{list_of_na_rle()} to an existing ggplot}
\usage{
layer_na_rle(var, data, ...)
}
\arguments{
\item{var}{A variable that contains \code{list_of_na_rle()}.}

\item{data}{A data frame.}

\item{...}{Passed to \code{geom_rect()}.}
}
\description{
Add a layer of \code{list_of_na_rle()} to an existing ggplot
}
\examples{
if (!requireNamespace("nycflights13", quietly = TRUE)) {
  stop("Please install the nycflights13 package to run these following examples.")
}
library(dplyr, warn.conflicts = FALSE)
weather <- nycflights13::weather
na_runs_wind <- weather \%>\% 
  group_by(origin) \%>\% 
  summarise_at(
    vars(contains("wind")), 
    ~ list_of_na_rle(., index_by = time_hour)
  )

library(ggplot2)
ggplot(weather, aes(x = time_hour, y = wind_gust)) +
  geom_line() +
  layer_na_rle(wind_gust, data = na_runs_wind, fill = "red", alpha = .5) +
  facet_grid(origin ~ .)
}
