% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na.R
\name{na_starts_with}
\alias{na_starts_with}
\alias{na_ends_with}
\title{Helpers for missing values}
\usage{
na_starts_with(x)

na_ends_with(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Helpers for missing values
}
\examples{
na_starts_with(c(rep(NA, 4), 10:1, NA))
na_ends_with(c(rep(NA, 4), 10:1, NA))
}
