% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na-rle.R
\name{na_rle_expand}
\alias{na_rle_expand}
\alias{na_rle_table}
\title{Expand and count run length encoding <\code{NA}>}
\usage{
na_rle_expand(x, ...)

na_rle_table(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=na_rle]{na_rle()}} or \code{\link[=list_of_na_rle]{list_of_na_rle()}}.}

\item{...}{A name-value pair adding a new column as identifiers,
if \code{\link[=list_of_na_rle]{list_of_na_rle()}}; otherwise ignored.}
}
\description{
Expand and count run length encoding <\code{NA}>
}
\examples{
(x <- na_rle(c(1, NA, NA, 4:7, NA, NA, 10:15, NA)))
na_rle_expand(x)
na_rle_table(x)
}
