% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mists-package.R
\docType{package}
\name{mists-package}
\alias{mists-package}
\title{mists: missingess of time series}
\description{
Provides 1d, 2d, and visual methods for exploring,
    polishing, and visualising temporal missingness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.earo.me/mists}
  \item Report bugs at \url{https://github.com/earowang/mists/issues}
}

}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com} (0000-0001-6448-5260)

Other contributors:
\itemize{
  \item Di Cook (0000-0002-3813-7155) [thesis advisor]
  \item Rob Hyndman (0000-0002-2140-5352) [thesis advisor]
}

}
