% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.mists_rle_na}
\alias{autoplot.mists_rle_na}
\alias{na_rle_spinogram}
\title{Autoplot and spinogram for run length encoding <\code{NA}>}
\usage{
\method{autoplot}{mists_rle_na}(object, ...)

na_rle_spinogram(x, y = NULL)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{...}{other arguments passed to specific methods}

\item{x, y}{Objects returned by \code{\link[=na_rle]{na_rle()}}.}
}
\description{
Autoplot and spinogram for run length encoding <\code{NA}>
}
\examples{
if (!requireNamespace("nycflights13", quietly = TRUE)) {
  stop("Please install the nycflights13 package to run these following examples.")
}
library(dplyr, warn.conflicts = FALSE)
na_runs_wind <- nycflights13::weather \%>\% 
  group_by(origin) \%>\% 
  summarise_at(vars(contains("wind")), ~ list_of_na_rle(., time_hour))

autoplot(na_runs_wind$wind_dir[[1]])
autoplot(na_runs_wind$wind_dir, na_runs_wind$origin)
autoplot(
  vctrs::as_list_of(na_runs_wind$wind_dir[[2]], na_runs_wind$wind_gust[[2]]),
  y = paste(c("wind_dir", "wind_gust"), "JFK", sep = "@"),
  size = 2.5
)
na_rle_spinogram(na_runs_wind$wind_dir[[1]])
na_rle_spinogram(na_runs_wind$wind_dir[[1]], na_runs_wind$wind_gust[[1]])
na_rle_spinogram(na_runs_wind$wind_dir[[1]], na_runs_wind$wind_dir[[3]])
}
