% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_auto}
\alias{na_polish_auto}
\alias{na_polish_auto_trace}
\title{Automate missing data polishing for tsibble}
\usage{
na_polish_auto(data, cutoff, tol = 0.1, quiet = FALSE)

na_polish_auto_trace(data, cutoff, tol = 0.1, quiet = FALSE)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A numeric between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}

\item{tol}{A tolerance value near zero as stopping rule. It compares to
a metric defined as \code{prop_na * prop_removed} to be minimised. See
\code{\link[=na_polish_metrics]{na_polish_metrics()}} for details.}

\item{quiet}{If \code{FALSE}, report metrics while automatically polishing, and
requires the "cliapp" package to be installed.}
}
\description{
It is an iterative process for minimising a metric until a tolerance value.
\itemize{
\item \code{na_polish_auto()} returns the polished data.
\item \code{na_polish_auto_trace()} returns a metric tibble for documenting the steps.
}
}
\examples{
\dontrun{
wdi_ts <- tsibble::as_tsibble(wdi, key = country_code, index = year)
wdi_after <- na_polish_auto(wdi_ts, cutoff = .8)
na_polish_metrics(wdi_ts, wdi_after)
}
}
