% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_auto}
\alias{na_polish_auto}
\alias{na_polish_autotrace}
\alias{na_polish_funs}
\title{Automate missing data polishing for tsibble}
\usage{
na_polish_auto(data, cutoff, tol = 0.1, funs = na_polish_funs(),
  quiet = FALSE)

na_polish_autotrace(data, cutoff, tol = 0.1, funs = na_polish_funs(),
  quiet = FALSE)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A numeric between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}

\item{tol}{A tolerance value close or equal to zero as stopping rule. It
compares to the loss defined as \code{prop_na * prop_removed} to be minimised.
See \code{\link[=na_polish_metrics]{na_polish_metrics()}} for details.}

\item{funs}{A list of \code{na_polish_*()} functions to go through.}

\item{quiet}{If \code{FALSE}, report metrics at each step and pass of the polishing
process. It requires the "cliapp" package to be installed.}
}
\description{
It is an iterative process for minimising the loss until a tolerance value.
\itemize{
\item \code{na_polish_auto()} returns the polished data.
\item \code{na_polish_autotrace()} returns a tibble for documenting the steps and metrics.
}
}
\examples{
\dontrun{
wdi_ts <- tsibble::as_tsibble(wdi, key = country_code, index = year)
wdi_after <- na_polish_auto(wdi_ts, cutoff = .8)
na_polish_metrics(wdi_ts, wdi_after)

# Trace down `na_polish_auto()`
na_polish_autotrace(wdi_ts, cutoff = .8, quiet = TRUE)
}
}
\seealso{
Other missing value polishing functions: \code{\link{na_polish_measures}},
  \code{\link{na_polish_metrics}}
}
\concept{missing value polishing functions}
\keyword{internal}
