% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_measures}
\alias{na_polish_measures}
\title{Data polishing for missing values in tsibble}
\usage{
na_polish_measures(data, cutoff)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A number between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}
}
\description{
Data polishing for missing values in tsibble
}
\details{
The proportion of overall missings is defined as the number of \code{NA} divided
by the number of \strong{measurements} (i.e. excluding key and index).
}
