% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_metrics}
\alias{na_polish_metrics}
\title{Metrics for missing data polishing}
\usage{
na_polish_metrics(before, after)
}
\arguments{
\item{before, after}{A tsibble.}
}
\description{
Metrics for missing data polishing
}
