% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na-rle.R
\name{na_rle_cut}
\alias{na_rle_cut}
\title{Cut and aggregate run length encoding <\code{NA}>}
\usage{
na_rle_cut(x, by, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=na_rle]{na_rle()}} or \code{\link[=list_of_na_rle]{list_of_na_rle()}}.}

\item{by}{A function applied to \code{indices}, such as tsibble's period functions
and lubridate's friends.}

\item{...}{A name-value pair adding a new column as identifiers,
if \code{\link[=list_of_na_rle]{list_of_na_rle()}}; otherwise ignored.}
}
\value{
A tibble contains:
\itemize{
\item \code{indices}: aggregated indices.
\item \code{n_run}: the number of unique runs for each \code{by}.
\item \code{n_na}: the total number of \code{NA}s for each \code{by}.
}
}
\description{
Cut and aggregate run length encoding <\code{NA}>
}
\examples{
if (!requireNamespace("nycflights13", quietly = TRUE)) {
  stop("Please install the nycflights13 package to run these following examples.")
}
library(dplyr, warn.conflicts = FALSE)
na_runs_wind <- nycflights13::weather \%>\% 
  group_by(origin) \%>\% 
  summarise_at(vars(contains("wind")), ~ list_of_na_rle(., time_hour))
na_rle_cut(na_runs_wind$wind_gust, by = tsibble::yearmonth)
na_rle_cut(na_runs_wind$wind_gust, by = tsibble::yearmonth,
  origin = na_runs_wind$origin)
}
