% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na-rle.R
\name{na_rle_shift}
\alias{na_rle_shift}
\title{Shift run length encoding <\code{NA}>}
\usage{
na_rle_shift(x, n = 1L)
}
\arguments{
\item{x}{An object returned by \code{\link[=na_rle]{na_rle()}} or \code{\link[=list_of_na_rle]{list_of_na_rle()}}.}

\item{n}{A positive integer shifts the starting indices forward, and negative
backward.}
}
\description{
Shift run length encoding <\code{NA}>
}
\examples{
(x <- na_rle(c(1, NA, NA, 4:7, NA, NA, 10:15, NA)))
na_rle_shift(x, n = 2)
na_rle_shift(x, n = -5)
}
