% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{as_list_of.rle_na}
\alias{vec_cast.list.rle_na}
\alias{vec_math.rle_na}
\alias{vec_math.list_of_rle_na}
\alias{obj_print_data.rle_na}
\alias{obj_print_header.rle_na}
\alias{vec_ptype_full.rle_na}
\alias{vec_ptype_abbr.rle_na}
\alias{vec_ptype_abbr.list_of_rle_na}
\alias{vec_ptype_full.list_of_rle_na}
\title{vctrs compatibility functions}
\usage{
\method{as_list_of}{rle_na}(x, ...)

\method{vec_cast.list}{rle_na}(x, to, ...)

\method{vec_math}{rle_na}(.fn, x, ...)

\method{vec_math}{list_of_rle_na}(.fn, x, ...)

\method{obj_print_data}{rle_na}(x, ...)

\method{obj_print_header}{rle_na}(x, ...)

\method{vec_ptype_full}{rle_na}(x)

\method{vec_ptype_abbr}{rle_na}(x)

\method{vec_ptype_abbr}{list_of_rle_na}(x)

\method{vec_ptype_full}{list_of_rle_na}(x)
}
\arguments{
\item{x}{Objects.}

\item{...}{Used to pass along error message information.}

\item{to}{Type to cast to.}
}
\description{
These functions are the extensions that allow mists objects to
work with vctrs.
}
\keyword{internal}
