% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdi}
\alias{wdi}
\title{World development indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10850 rows and 57 columns.}
\usage{
wdi
}
\description{
World development indicators
}
\references{
\href{https://databank.worldbank.org/data/source/world-development-indicators/}{The world bank databank}
}
\keyword{datasets}
