% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{count_na}
\alias{count_na}
\title{Count explicit NA}
\usage{
count_na(.data, ...)
}
\arguments{
\item{.data}{A `tbl_ts`.}

\item{...}{A bare variable that may contain `NA`.}
}
\value{
A tibble contains:
* the "key" of the `tbl_ts`
* ".from": the starting time point of `NA`
* ".to": the ending time point of `NA`
* ".n": the number of `NA` during the time period
}
\description{
Count explicit NA
}
\examples{
pedestrian \%>\% 
  fill_gaps(.full = TRUE) \%>\% 
  count_na(Count)
}
