% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na.R
\name{na_starts_with}
\alias{na_starts_with}
\alias{na_ends_with}
\alias{na_elsewhere}
\title{Handlers for missing values}
\usage{
na_starts_with(x)

na_ends_with(x)

na_elsewhere(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Handlers for missing values
}
\examples{
x <- c(rep(NA, 4), 10:6, NA, 4:1, NA)
na_starts_with(x)
na_ends_with(x)
na_elsewhere(x)
}
