% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mists-package.R
\docType{package}
\name{mists-package}
\alias{mists-package}
\title{mists: missingess of time series}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com}

}
