% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mists-package.R
\docType{package}
\name{mists-package}
\alias{mists-package}
\title{mists: missingess of time series}
\description{
Provides 1d, 2d, and visual methods for exploring, polishing, and
    visualising temporal missingness.
}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com}

}
