% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{na_rle_rangeplot}
\alias{na_rle_rangeplot}
\alias{na_rle_spinoplot}
\alias{na_rle_spineplot}
\alias{autoplot.rle_na}
\title{Range plot and spinoplot for runs of missings}
\usage{
na_rle_rangeplot(data, x, y = NULL, ...)

na_rle_spinoplot(data, x, y = NULL, facets = NULL, ...)

\method{autoplot}{rle_na}(object, y = as.factor(1L), ...)
}
\arguments{
\item{data}{A data frame that contains \code{\link[=list_of_na_rle]{list_of_na_rle()}}.}

\item{x, y}{A bare variable contains \code{\link[=list_of_na_rle]{list_of_na_rle()}} mapped to x and y.}

\item{...}{\itemize{
\item \code{na_rle_rangeplot()}: passed to \code{geom_line()} and \code{geom_point()}.
\item \code{na_rle_spinoplot()}: passed to \code{facet_wrap()}.
}}

\item{facets}{A facetting variable.}

\item{object}{an object, whose class will determine the behaviour of autoplot}
}
\description{
\code{na_rle_spineplot()} is an alias of \code{na_rle_spinoplot()}.
}
\examples{
if (!requireNamespace("nycflights13", quietly = TRUE)) {
  stop("Please install the nycflights13 package to run these following examples.")
}
library(dplyr, warn.conflicts = FALSE)
na_runs_wind <- nycflights13::weather \%>\% 
  group_by(origin) \%>\% 
  summarise_at(vars(contains("wind")), ~ list_of_na_rle(., time_hour))

na_runs_wind \%>\% 
  na_rle_rangeplot(wind_dir, origin, shape = 4)
# autoplot() method
autoplot(na_runs_wind$wind_gust)
na_runs_wind \%>\%
  na_rle_spinoplot(x = wind_dir, facets = origin)
na_runs_wind \%>\%
  na_rle_spinoplot(x = wind_dir, y = wind_gust, facets = origin)
}
