% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.mists_rle_na}
\alias{autoplot.mists_rle_na}
\alias{na_rle_spinogram}
\title{Autoplot and spinogram for run length encoding <\code{NA}>}
\usage{
\method{autoplot}{mists_rle_na}(object, ...)

na_rle_spinogram(x, y = NULL)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{...}{other arguments passed to specific methods}

\item{x, y}{Objects returned by \code{\link[=na_rle]{na_rle()}}.}
}
\description{
Autoplot and spinogram for run length encoding <\code{NA}>
}
\examples{
library(dplyr, warn.conflicts = FALSE)
na_runs_wind <- nycflights13::weather \%>\% 
  group_by(origin) \%>\% 
  summarise_at(vars(contains("wind")), ~ list_of_na_rle(., time_hour))

autoplot(na_runs_wind$wind_dir[[1]])
autoplot(na_runs_wind$wind_dir, na_runs_wind$origin)
na_rle_spinogram(na_runs_wind$wind_dir[[1]])
na_rle_spinogram(na_runs_wind$wind_dir[[1]], na_runs_wind$wind_gust[[1]])
na_rle_spinogram(na_runs_wind$wind_dir[[1]], na_runs_wind$wind_dir[[3]])
}
