% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_auto}
\alias{na_polish_auto}
\alias{na_polish_auto_trace}
\title{Automatic data polishing for missing values in tsibble}
\usage{
na_polish_auto(data, cutoff, tol = 0.1, quiet = FALSE)

na_polish_auto_trace(data, cutoff, tol = 0.1, quiet = FALSE)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A numeric between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}

\item{tol}{A tolerence value as stopping rule.}

\item{quiet}{Report metrics along the way of the automatic polishing.}
}
\description{
Automatic data polishing for missing values in tsibble
}
\examples{
library(tsibble, warn.conflicts = FALSE)
wdi_ts <- as_tsibble(wdi, key = country_code, index = year)
wdi_after <- na_polish_auto(wdi_ts, cutoff = .8)
na_polish_metrics(wdi_ts, wdi_after)
}
