% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_measures}
\alias{na_polish_measures}
\alias{na_polish_key}
\alias{na_polish_index}
\alias{na_polish_index2}
\title{Missing data polishing for tsibble}
\usage{
na_polish_measures(data, cutoff)

na_polish_key(data, cutoff)

na_polish_index(data, cutoff, na_fun = na_starts_with)

na_polish_index2(data, cutoff)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A numeric between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}

\item{na_fun}{Either \code{\link{na_starts_with}} or \code{\link{na_ends_with}}.}
}
\description{
If the proportion of overall missings is less than the cutoff,
\itemize{
\item \code{na_polish_measures()} removes columns or observations.
\item \code{na_polish_key()} polishes data by rows or observations, removing the whole
rows of key series.
\item \code{na_polish_index()} polishes data by rows or observations, removing either
the starting or the ending \code{NA} blocks (if any) within each key series.
\item \code{na_polish_index2()} polishes data by rows or observations, removing the
ending \code{NA} blocks (if any) within each key series. It is a shortcut of
\code{na_polish_index(na_fun = na_ends_with)}.
}
}
\details{
The proportion of overall missings is defined as the number of \code{NA} divided
by the number of \strong{measurements} (i.e. excluding key and index).
}
\examples{
wdi_ts <- tsibble::as_tsibble(wdi, key = country_code, index = year)
(wdi_cols <- na_polish_measures(wdi_ts, cutoff = .7))
# columns removed
setdiff(names(wdi_ts), names(wdi_cols))
na_polish_key(wdi_ts, cutoff = .7)
na_polish_index(wdi_ts, cutoff = .7)
na_polish_index2(wdi_ts, cutoff = .7)
}
