% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{na_polish_measures}
\alias{na_polish_measures}
\alias{na_polish_key}
\alias{na_polish_index}
\alias{na_polish_index2}
\title{Missing data polishing for tsibble}
\usage{
na_polish_measures(data, cutoff)

na_polish_key(data, cutoff)

na_polish_index(data, cutoff, na_fun = na_starts_with)

na_polish_index2(data, cutoff)
}
\arguments{
\item{data}{A tsibble.}

\item{cutoff}{A numeric between 0 and 1. Rows/cols will be kept, if the
proportion of overall missings is less than the cutoff.}

\item{na_fun}{Either \code{\link{na_starts_with}} or \code{\link{na_ends_with}}.}
}
\description{
Missing data polishing for tsibble
}
\details{
The proportion of overall missings is defined as the number of \code{NA} divided
by the number of \strong{measurements} (i.e. excluding key and index).
}
