% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na-rle.R
\name{set-op}
\alias{set-op}
\alias{intersect.mists_rle_na}
\alias{union.mists_rle_na}
\alias{setdiff.mists_rle_na}
\title{Set operations for run length encoding <\code{NA}>}
\usage{
\method{intersect}{mists_rle_na}(x, y, ...)

\method{union}{mists_rle_na}(x, y, ...)

\method{setdiff}{mists_rle_na}(x, y, ...)
}
\arguments{
\item{x, y}{Objects returned by \code{\link[=na_rle]{na_rle()}} or \code{\link[=list_of_na_rle]{list_of_na_rle()}}.}

\item{...}{other arguments passed on to methods}
}
\description{
Set operations for run length encoding <\code{NA}>
}
\examples{
(x <- na_rle(c(1, NA, NA, 4:7, NA, NA, 10:15, NA)))
(y <- na_rle(c(10, NA, NA, NA, 6:3, NA, 1)))
intersect(x, y)
union(x, y)
setdiff(x, y)
setdiff(y, x)
}
