% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mists-rle-na.R
\name{na_rle}
\alias{na_rle}
\alias{na_rle_lengths}
\alias{na_rle_indices}
\alias{list_of_na_rle}
\alias{as_list_of.mists_rle_na}
\title{Run lengths encoding for missing values (\code{NA})}
\usage{
na_rle(x = double(), index_by = seq_along(x), interval = NULL)

na_rle_lengths(x)

na_rle_indices(x)

list_of_na_rle(x = double(), index_by = seq_along(x),
  interval = NULL)

\method{as_list_of}{mists_rle_na}(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{index_by}{A vector of the same length as \code{x}.}

\item{interval}{if \code{NULL}, determined by the greatest common denominator;
otherwise a supplied "interval" class. See \code{?tsibble::tsibble} for details.}

\item{...}{Vectors to coerce.}
}
\value{
A named list of
\itemize{
\item \code{lengths}: the lengths of \code{NA} runs
\item \code{indices}: the starting indices of runs
}
}
\description{
Run lengths encoding for missing values (\code{NA})
}
\section{Mathematical operations}{

Many math operations can be applied to objects returned from \code{na_rle()} and
\code{list_of_na_rle()}, regarding the \emph{lengths} of runs.
\itemize{
\item \code{sum()}: the total number of \code{NA} over all runs.
\item \code{mean()}: the average \code{NA}s per run.
\item \code{min()} & \code{max()}: the minimum and maximum of runs.
\item \code{median()} & \code{quantile()}
}
}

\examples{
df <- data.frame(year = 2000:2019, temp = sample(0:30, size = 10))
df[c(1, 6, 13:16, 19), "temp"] <- NA

na_rle(df$temp) # indexed by the default positions
(x <- na_rle(df$temp, index_by = df$year)) # indexed by a variable

na_rle_lengths(x)
na_rle_indices(x)

length(x) # the number of runs
sum(x) # the total number of `NA`
range(x) # min & max runs

library(dplyr, warn.conflicts = FALSE)
# list_of_na_rle() is useful when working with tabular data
na_rle_df <- as_tibble(df) \%>\%
  summarise(na_runs = list_of_na_rle(temp, year))
na_rle_df

length(na_rle_df$na_runs)
sum(na_rle_df$na_runs)
range(na_rle_df$na_runs)
}
