% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mists-rle-na.R
\name{na_rle}
\alias{na_rle}
\alias{na_rle_lengths}
\alias{na_rle_values}
\alias{list_of_na_rle}
\title{Run lengths encoding for missing values (\code{NA})}
\usage{
na_rle(x = double(), index_by = seq_along(x), interval = NULL)

na_rle_lengths(x)

na_rle_values(x)

list_of_na_rle(x = double(), index_by = seq_along(x),
  interval = NULL)
}
\arguments{
\item{x}{A vector.}

\item{index_by}{A vector of the same length as \code{x}.}

\item{interval}{if \code{NULL}, determined by the greatest common denominator;
otherwise a supplied "interval" class.}
}
\value{
A named list of
\itemize{
\item \code{lengths}: the lengths of \code{NA} runs
\item \code{values}: the starting indices of runs
}
}
\description{
Run lengths encoding for missing values (\code{NA})
}
\examples{
df <- data.frame(year = 2000:2019, temp = sample(0:30, size = 10))
df[c(1, 6, 13:16, 19), "temp"] <- NA

na_rle(df$temp) # indexed by the default positions
(x <- na_rle(df$temp, index_by = df$year)) # indexed by a variable

na_rle_lengths(x)
na_rle_values(x)

library(dplyr, warn.conflicts = FALSE)
# list_of_na_rle() is useful when working with tabular data
as_tibble(df) \%>\% 
  summarise(na_runs = list_of_na_rle(temp, year))
}
