% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-na-rle.R
\name{na_rle_cut}
\alias{na_rle_cut}
\title{Cut and aggregate run length encoding <\code{NA}>}
\usage{
na_rle_cut(x, by)
}
\arguments{
\item{x}{An object returned by \code{\link[=na_rle]{na_rle()}}.}

\item{by}{A function applied to \code{indices}, such as tsibble's period functions
and lubridate's friends.}
}
\value{
A tibble contains:
\itemize{
\item \code{indices}: aggregated indices.
\item \code{n_run}: the number of unique runs for each \code{by}.
\item \code{n_na}: the total number of \code{NA}s for each \code{by}.
}
}
\description{
Cut and aggregate run length encoding <\code{NA}>
}
\examples{
if (!requireNamespace("nycflights13", quietly = TRUE)) {
  stop("Please install the nycflights13 package to run these following examples.")
}
if (!requireNamespace("tidyr", quietly = TRUE)) {
  stop("Please install the tidyr package to run these following examples.")
}
library(dplyr, warn.conflicts = FALSE)
nycflights13::weather \%>\% 
  group_by(origin) \%>\% 
  summarise(wind_gust_na = list_of_na_rle(wind_gust, time_hour)) \%>\% 
  mutate(wind_gust_na_cut = na_rle_cut(wind_gust_na, by = tsibble::yearmonth)) \%>\% 
  tidyr::unnest(cols = wind_gust_na_cut)
}
