% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{as_list_of.mists_rle_na}
\alias{vec_cast.list.mists_rle_na}
\alias{obj_print_data.mists_rle_na}
\alias{vec_ptype_full.mists_rle_na}
\alias{vec_ptype_abbr.mists_rle_na}
\alias{vec_ptype_abbr.mists_list_of_rle_na}
\alias{vec_ptype_full.mists_list_of_rle_na}
\title{vctrs compatibility functions}
\usage{
\method{as_list_of}{mists_rle_na}(x, ...)

\method{vec_cast.list}{mists_rle_na}(x, to)

\method{obj_print_data}{mists_rle_na}(x, ...)

\method{vec_ptype_full}{mists_rle_na}(x)

\method{vec_ptype_abbr}{mists_rle_na}(x)

\method{vec_ptype_abbr}{mists_list_of_rle_na}(x)

\method{vec_ptype_full}{mists_list_of_rle_na}(x)
}
\arguments{
\item{x}{Objects.}

\item{...}{Used to pass along error message information.}

\item{to}{Type to cast to.}
}
\description{
These functions are the extensions that allow mists objects to
work with vctrs.
}
